Test to illustrate problem that adding fade screws up static buffer len
To loop the most recent part of the static buffer requires more
recent samples on the end to add the fade out.
This means we would need to tick the static buffer for more time,
as the fade can be changed dynamically by the client (?) it would
need an amount of samples equal to the longest possible fade.
This would also screw up the various offset settings when using
the static buffer, everything will need an extra fade len applied.